pi.html is complete!!!!!!
First, there is a black background. To make pacman, I made a circle and then a triangle cut out the same color as the background. Then, I put in the 3.1415 going into pacman's mouth. But since the 3.1415 goes on top of pacman instead of in him, I had to make two triangles and a rectangle to cover it up. Then I made yum.